add Game header and implementation

+++ Headers/Vertex.h
@@ -8,6 +8,6 @@ struct Vertex
     float animated_y;
     int bone_id1;
     int bone_id2;
-    float bone_weight1;
-    float bone_weight2;
+    double bone_weight1;
+    double bone_weight2;
 };
